issue #115 rename : model -> model/socket/chat
